Managers updates + breed manager + diverse small changes

+++ src/main/kotlin/fr/lewon/dofus/bot/core/model/hunt/DofusPointOfInterest.kt
@@ -1,6 +1,6 @@
 package fr.lewon.dofus.bot.core.model.hunt
 
 data class DofusPointOfInterest(
-    val id: Double,
+    val id: Int,
     val label: String
 )
